# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sawtooth Brewery Keg Ordering API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Beer {
  id: number;
  name: string;
  kegSize: string;
  price: number;
  available: boolean;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBeerBody {
  name: string;
  kegSize: string;
  price: number;
  available: boolean;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateBeerBody {
  name?: string;
  kegSize?: string;
  price?: number;
  available?: boolean;
  /** @nullable */
  notes?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  authorized: 'authorized',
  captured: 'captured',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  pickupDate: string;
  pickupTime: string;
  beerId: number;
  beerName: string;
  kegSize: string;
  quantity: number;
  depositAmount: number;
  partyTapNeeded: boolean;
  co2Needed: boolean;
  /** @nullable */
  notes: string | null;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  cloverPaymentId: string | null;
  totalAmount: number;
  /** @nullable */
  paymentError: string | null;
  createdAt: string;
  updatedAt: string;
}

export type OrderDetailStatus = typeof OrderDetailStatus[keyof typeof OrderDetailStatus];


export const OrderDetailStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export type OrderDetailPaymentStatus = typeof OrderDetailPaymentStatus[keyof typeof OrderDetailPaymentStatus];


export const OrderDetailPaymentStatus = {
  pending: 'pending',
  authorized: 'authorized',
  captured: 'captured',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface KegReceipt {
  id: number;
  orderId: number;
  /** @nullable */
  dateOfSale: string | null;
  /** @nullable */
  dateOfReturn: string | null;
  /** @nullable */
  purchaserName: string | null;
  /** @nullable */
  purchaserDob: string | null;
  /** @nullable */
  purchaserPhone: string | null;
  /** @nullable */
  consumptionLocation: string | null;
  /** @nullable */
  consumptionTime: string | null;
  /** @nullable */
  consumptionDate: string | null;
  /** @nullable */
  validIdNumber: string | null;
  /** @nullable */
  vehicleYear: string | null;
  /** @nullable */
  vehicleMake: string | null;
  /** @nullable */
  vehicleColor: string | null;
  /** @nullable */
  vehiclePlate: string | null;
  /** @nullable */
  customerSignature: string | null;
  /** @nullable */
  signedAt: string | null;
  /** @nullable */
  staffName: string | null;
  /** @nullable */
  draftSystemNumber: string | null;
  /** @nullable */
  co2TankNeeded: boolean | null;
  /** @nullable */
  co2RegulatorNumber: string | null;
  /** @nullable */
  trashCanNumbers: string | null;
  /** @nullable */
  kegBrand: string | null;
  /** @nullable */
  kegSize: string | null;
  /** @nullable */
  kegIdNumbers: string | null;
  completed: boolean;
  createdAt: string;
  updatedAt: string;
}

export type PickupStatus = typeof PickupStatus[keyof typeof PickupStatus];


export const PickupStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Pickup {
  id: number;
  orderId: number;
  customerName: string;
  pickupDate: string;
  pickupTime: string;
  beerName: string;
  kegSize: string;
  quantity: number;
  status: PickupStatus;
  createdAt: string;
}

export interface OrderDetail {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  pickupDate: string;
  pickupTime: string;
  beerId: number;
  beerName: string;
  kegSize: string;
  quantity: number;
  depositAmount: number;
  partyTapNeeded: boolean;
  co2Needed: boolean;
  /** @nullable */
  notes: string | null;
  status: OrderDetailStatus;
  paymentStatus: OrderDetailPaymentStatus;
  /** @nullable */
  cloverPaymentId: string | null;
  totalAmount: number;
  /** @nullable */
  paymentError: string | null;
  receipt: KegReceipt;
  pickup: Pickup;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  pickupDate: string;
  pickupTime: string;
  beerId: number;
  quantity: number;
  partyTapNeeded: boolean;
  co2Needed: boolean;
  /** @nullable */
  notes?: string | null;
  cloverPaymentToken: string;
  idempotencyKey: string;
}

export type UpdateOrderBodyStatus = typeof UpdateOrderBodyStatus[keyof typeof UpdateOrderBodyStatus];


export const UpdateOrderBodyStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface UpdateOrderBody {
  status?: UpdateOrderBodyStatus;
  /** @nullable */
  notes?: string | null;
  pickupDate?: string;
  pickupTime?: string;
}

export interface AuthorizePaymentBody {
  amount: number;
  source: string;
  idempotencyKey: string;
  /** @nullable */
  orderId?: number | null;
}

export interface PaymentAuthorization {
  id: number;
  cloverPaymentId: string;
  amount: number;
  status: string;
  idempotencyKey: string;
  /** @nullable */
  orderId: number | null;
  createdAt: string;
}

export interface CreateKegReceiptBody {
  /** @nullable */
  dateOfSale?: string | null;
  /** @nullable */
  dateOfReturn?: string | null;
  /** @nullable */
  purchaserName?: string | null;
  /** @nullable */
  purchaserDob?: string | null;
  /** @nullable */
  purchaserPhone?: string | null;
  /** @nullable */
  consumptionLocation?: string | null;
  /** @nullable */
  consumptionTime?: string | null;
  /** @nullable */
  consumptionDate?: string | null;
  /** @nullable */
  validIdNumber?: string | null;
  /** @nullable */
  vehicleYear?: string | null;
  /** @nullable */
  vehicleMake?: string | null;
  /** @nullable */
  vehicleColor?: string | null;
  /** @nullable */
  vehiclePlate?: string | null;
  /** @nullable */
  staffName?: string | null;
  /** @nullable */
  draftSystemNumber?: string | null;
  /** @nullable */
  co2TankNeeded?: boolean | null;
  /** @nullable */
  co2RegulatorNumber?: string | null;
  /** @nullable */
  trashCanNumbers?: string | null;
  /** @nullable */
  kegBrand?: string | null;
  /** @nullable */
  kegSize?: string | null;
  /** @nullable */
  kegIdNumbers?: string | null;
}

export interface SaveSignatureBody {
  customerSignature: string;
  signedAt: string;
}

export interface DashboardSummary {
  totalOrders: number;
  pendingOrders: number;
  confirmedOrders: number;
  completedOrders: number;
  totalRevenue: number;
  pendingRevenue: number;
  upcomingPickups: number;
  todayPickups: number;
}

export type ListBeersParams = {
availableOnly?: boolean;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export type ListPickupsParams = {
startDate?: string;
endDate?: string;
};

export type GetRecentOrdersParams = {
limit?: number;
};