# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sawtooth Brewery Keg Ordering API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /receipts/{orderId}:
    get:
      operationId: getKegReceipt
      tags: [receipts]
      summary: Get Idaho keg receipt for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Keg receipt form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KegReceipt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrUpdateKegReceipt
      tags: [receipts]
      summary: Create or update Idaho keg receipt for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKegReceiptBody"
      responses:
        "200":
          description: Keg receipt saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KegReceipt"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /receipts/{orderId}/signature:
    post:
      operationId: saveReceiptSignature
      tags: [receipts]
      summary: Save customer signature for a receipt
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSignatureBody"
      responses:
        "200":
          description: Signature saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KegReceipt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Dashboard summary (wow endpoint)
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    KegReceipt:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        dateOfSale:
          type: ["string", "null"]
        dateOfReturn:
          type: ["string", "null"]
        purchaserName:
          type: ["string", "null"]
        purchaserDob:
          type: ["string", "null"]
        purchaserPhone:
          type: ["string", "null"]
        consumptionLocation:
          type: ["string", "null"]
        consumptionTime:
          type: ["string", "null"]
        consumptionDate:
          type: ["string", "null"]
        validIdNumber:
          type: ["string", "null"]
        vehicleYear:
          type: ["string", "null"]
        vehicleMake:
          type: ["string", "null"]
        vehicleColor:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        customerSignature:
          type: ["string", "null"]
        signedAt:
          type: ["string", "null"]
        staffName:
          type: ["string", "null"]
        draftSystemNumber:
          type: ["string", "null"]
        co2TankNeeded:
          type: ["boolean", "null"]
        co2RegulatorNumber:
          type: ["string", "null"]
        trashCanNumbers:
          type: ["string", "null"]
        kegBrand:
          type: ["string", "null"]
        kegSize:
          type: ["string", "null"]
        kegIdNumbers:
          type: ["string", "null"]
        completed:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderId
        - dateOfSale
        - dateOfReturn
        - purchaserName
        - purchaserDob
        - purchaserPhone
        - consumptionLocation
        - consumptionTime
        - consumptionDate
        - validIdNumber
        - vehicleYear
        - vehicleMake
        - vehicleColor
        - vehiclePlate
        - customerSignature
        - signedAt
        - staffName
        - draftSystemNumber
        - co2TankNeeded
        - co2RegulatorNumber
        - trashCanNumbers
        - kegBrand
        - kegSize
        - kegIdNumbers
        - completed
        - createdAt
        - updatedAt
    CreateKegReceiptBody:
      type: object
      properties:
        dateOfSale:
          type: ["string", "null"]
        dateOfReturn:
          type: ["string", "null"]
        purchaserName:
          type: ["string", "null"]
        purchaserDob:
          type: ["string", "null"]
        purchaserPhone:
          type: ["string", "null"]
        consumptionLocation:
          type: ["string", "null"]
        consumptionTime:
          type: ["string", "null"]
        consumptionDate:
          type: ["string", "null"]
        validIdNumber:
          type: ["string", "null"]
        vehicleYear:
          type: ["string", "null"]
        vehicleMake:
          type: ["string", "null"]
        vehicleColor:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        staffName:
          type: ["string", "null"]
        draftSystemNumber:
          type: ["string", "null"]
        co2TankNeeded:
          type: ["boolean", "null"]
        co2RegulatorNumber:
          type: ["string", "null"]
        trashCanNumbers:
          type: ["string", "null"]
        kegBrand:
          type: ["string", "null"]
        kegSize:
          type: ["string", "null"]
        kegIdNumbers:
          type: ["string", "null"]
    SaveSignatureBody:
      type: object
      properties:
        customerSignature:
          type: string
        signedAt:
          type: string
      required:
        - customerSignature
        - signedAt
...